cilium: ipsec, add route helpers to query for encrypt rules
IPsec encryption uses rules to direct packets that need encryption to
a routing table with necessary routes to hit local xfrm stack for
encrypt and decrypt operations. We are going to add some unit tests
for this but need a convienent way to lookup if rules exist or are
missing.

This patch adds a set of helpers to query rule state easily without
importing lots of other things such us linux_defaults and ipsec.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>